underground-model/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Cors==3.0.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.0
MarkupSafe==1.1.0
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pytest==4.1.1
requests==2.21.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
Werkzeug==0.14.1
